refactor: moved test constants to new file
