+++ __tests__/integration/utils/wallet-helper.js
@@ -1,5 +1,6 @@
 import { loggers } from './logger.util';
-import { TestUtils, WALLET_CONSTANTS } from './test-utils-integration';
+import { TestUtils } from './test-utils-integration';
+import { WALLET_CONSTANTS } from '../configuration/test-constants';
 import { WalletBenchmarkUtil } from './benchmark/wallet-benchmark.util';
 import { TxTimeHelper } from './benchmark/tx-time.helper';
 import { precalculationHelpers } from '../../../src/helpers/wallet-precalculation.helper';
